드롭다운 스타일 분리 및 오류 처리 추가
